refactor: 매핑 메서드 Mapper.kt 에서 각 Model의 메서드로 이동
