+++ android/app/src/main/java/com/woowacourse/ody/data/remote/core/entity/meeting/response/MeetingResponse.kt
@@ -2,6 +2,10 @@ package com.woowacourse.ody.data.remote.core.entity.meeting.response
 
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
+import com.woowacourse.ody.domain.model.Meeting
+import java.time.LocalDate
+import java.time.LocalTime
+import java.time.format.DateTimeFormatter
 
 @JsonClass(generateAdapter = true)
 data class MeetingResponse(
@@ -25,4 +29,25 @@ data class MeetingResponse(
     val mates: List<MateResponse>,
     @Json(name = "inviteCode")
     val inviteCode: String,
-)
+) {
+    fun toMeeting(): Meeting =
+        Meeting(
+            id = id,
+            name = name,
+            targetPosition = targetAddress,
+            meetingDate = date.parseToLocalDate(),
+            meetingTime = time.parseToLocalTime(),
+            mates = mates.map { it.toMate() },
+            inviteCode = inviteCode,
+        )
+}
+
+fun String.parseToLocalDate(): LocalDate {
+    val formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd")
+    return LocalDate.parse(this, formatter)
+}
+
+fun String.parseToLocalTime(): LocalTime {
+    val formatter = DateTimeFormatter.ofPattern("HH:mm:ss")
+    return LocalTime.parse(this, formatter)
+}
