+++ android/app/src/main/java/com/woowacourse/ody/data/remote/core/repository/DefaultMeetingRepository.kt
@@ -3,7 +3,6 @@ package com.woowacourse.ody.data.remote.core.repository
 import com.woowacourse.ody.data.remote.core.RetrofitClient
 import com.woowacourse.ody.data.remote.core.entity.meeting.request.MeetingRequest
 import com.woowacourse.ody.data.remote.core.entity.meeting.response.MeetingResponse
-import com.woowacourse.ody.data.remote.core.entity.toMeeting
 import com.woowacourse.ody.data.remote.core.service.MeetingService
 import com.woowacourse.ody.domain.model.Meeting
 import com.woowacourse.ody.domain.repository.ody.MeetingRepository
