+++ android/app/src/main/java/com/woowacourse/ody/data/remote/core/repository/DefaultNotificationLogRepository.kt
@@ -1,7 +1,6 @@
 package com.woowacourse.ody.data.remote.core.repository
 
 import com.woowacourse.ody.data.remote.core.RetrofitClient
-import com.woowacourse.ody.data.remote.core.entity.toNotificationList
 import com.woowacourse.ody.data.remote.core.service.NotificationService
 import com.woowacourse.ody.domain.model.NotificationLog
 import com.woowacourse.ody.domain.repository.ody.NotificationLogRepository
