+++ android/app/src/main/java/com/woowacourse/ody/domain/model/GeoLocation.kt
@@ -1,7 +1,11 @@
 package com.woowacourse.ody.domain.model
 
+import com.woowacourse.ody.presentation.address.model.GeoLocationUiModel
+
 class GeoLocation(
     val address: String,
     val longitude: String,
     val latitude: String,
-)
+) {
+    fun toGeoLocationUiModel(): GeoLocationUiModel = GeoLocationUiModel(address, longitude, latitude)
+}
