+++ android/app/src/main/java/com/woowacourse/ody/domain/model/Meeting.kt
@@ -1,7 +1,10 @@
 package com.woowacourse.ody.domain.model
 
+import com.woowacourse.ody.presentation.room.model.MeetingUiModel
 import java.time.LocalDate
+import java.time.LocalDateTime
 import java.time.LocalTime
+import java.time.format.DateTimeFormatter
 
 data class Meeting(
     val id: Long,
@@ -11,4 +14,20 @@ data class Meeting(
     val meetingTime: LocalTime,
     val mates: List<Mate>,
     val inviteCode: String,
-)
+) {
+    fun toMeetingUiModel(): MeetingUiModel {
+        return MeetingUiModel(
+            this.name,
+            this.targetPosition,
+            this.toMeetingDateTime(),
+            this.mates.map { it.nickname },
+            this.inviteCode,
+        )
+    }
+
+    private fun toMeetingDateTime(): String {
+        val dateTime = LocalDateTime.of(meetingDate, meetingTime)
+        val dateTimeFormatter = DateTimeFormatter.ofPattern("yyyy년 M월 d일 HH시 mm분")
+        return dateTime.format(dateTimeFormatter)
+    }
+}
