+++ android/app/src/main/java/com/woowacourse/ody/presentation/address/model/GeoLocationUiModel.kt
@@ -1,6 +1,7 @@
 package com.woowacourse.ody.presentation.address.model
 
 import android.os.Parcelable
+import com.woowacourse.ody.domain.model.GeoLocation
 import kotlinx.parcelize.Parcelize
 
 @Parcelize
@@ -8,4 +9,6 @@ class GeoLocationUiModel(
     val address: String,
     val longitude: String,
     val latitude: String,
-) : Parcelable
+) : Parcelable {
+    fun toGeoLocation(): GeoLocation = GeoLocation(address, longitude, latitude)
+}
